AutoPtr, amd64, consistent copy semantics
- amd64 fixes
- oz
  * AutoPtr pointer type that deletes the object it's pointing at on destruction
  * move semantics completely removed, copy semantics now consistent
  * polishing, fixes

+++ src/client/BasicAudio.cpp
@@ -41,7 +41,7 @@ namespace client
     {
       float dv = Math::sqrt( dyn->velocity.x*dyn->velocity.x +
                              dyn->velocity.y*dyn->velocity.y );
-      playContSound( samples[SND_FRICTING], dv, uint( &*dyn ), parent->obj );
+      playContSound( samples[SND_FRICTING], dv, uint( size_t( &*dyn ) ), parent->obj );
     }
 
     // events
